Create Pennar_PCA_FuzzyClustering_ANN_for_tutorial.m
This is the solution for a tutorial on PCA, Fuzzy Clustering and ANN model fitting that I was going to help deliver to trainees at IISc when I was a project assistant there. But unfortunately I fell sick with Chicken Pox and ended up reading Dostoevsky's 'Crime and Punishment' in an isolation ward.